Add responsive sidebar
Add map controls dynamic position

Fix sidebar in desktop

Add navigation menu to map header

Fix header width mobile

Fix zoom button position
