+++ src/components/map/controls/bookmark/index.tsx
@@ -56,7 +56,7 @@ export const BookmarkControl: FC = () => {
       <SheetTrigger className={CONTROL_BUTTON_STYLES.default}>
         <AiFillStar className={CONTROL_ICON_STYLES.default} />
       </SheetTrigger>
-      <SheetContent className="relative flex h-full flex-col space-y-6 bg-brand-500 bg-opacity-90 pl-10">
+      <SheetContent className="relative z-[55] flex h-full flex-col space-y-6 bg-brand-500 bg-opacity-90 sm:pl-10">
         <SheetHeader className="space-y-6">
           <SheetTitle className="text-2xl font-bold text-secondary-500">Bookmarks</SheetTitle>
           <div className="scroll-y-auto">
@@ -101,7 +101,7 @@ export const BookmarkControl: FC = () => {
         </SheetHeader>
 
         {!isInputVisible && (
-          <Button className="w-full" onClick={() => setInputVisibility(true)}>
+          <Button className="h-11 w-full sm:h-9" onClick={() => setInputVisibility(true)}>
             Bookmark current URL
           </Button>
         )}
