+++ src/components/ui/dialog.tsx
@@ -65,12 +65,12 @@ const DialogContent = forwardRef<
     <DialogPrimitive.Content
       ref={ref}
       className={cn(
-        'fixed left-[50%] top-[70px] z-50 flex max-h-[80%] w-full max-w-3xl translate-x-[-50%] flex-col bg-brand-400 shadow-lg duration-200 data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[state=closed]:slide-out-to-left-1/2 data-[state=closed]:slide-out-to-top-[48%] data-[state=open]:slide-in-from-left-1/2 data-[state=open]:slide-in-from-top-[48%]',
+        'fixed left-[50%] top-0 z-50 flex max-h-full w-full max-w-3xl translate-x-[-50%] flex-col bg-brand-400 shadow-lg duration-200 data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[state=closed]:slide-out-to-left-1/2 data-[state=closed]:slide-out-to-top-[48%] data-[state=open]:slide-in-from-left-1/2 data-[state=open]:slide-in-from-top-[48%] sm:top-[70px] sm:max-h-[80%]',
         className
       )}
       {...props}
     >
-      <ScrollArea className="grow p-10">{children}</ScrollArea>
+      <ScrollArea className="grow px-4 py-10 sm:px-10">{children}</ScrollArea>
     </DialogPrimitive.Content>
   </DialogPrimitive.Portal>
 ));
