+++ src/styles/globals.css
@@ -132,11 +132,11 @@
 }
 
 /* zoom control ropenlayers */
-.ol-zoom {
-  display: flex;
+.ol-zoom, .ol-zoom-open {
+  display: none;
   flex-direction: column !important;
-  position: absolute !important;
-  left: 553px !important;
+  /* position: relative; */
+  /* left: 16px !important; */
   width: 34px !important;
   height: 68px !important;
   align-items: center;
@@ -144,10 +144,39 @@
   border-radius: calc(var(--radius) - 4px) !important;
   background-color: hsl(210, 54%, 9%) !important;
   top: unset !important;
-  bottom: 90px !important;
+  bottom: 96px !important;
 }
 
-.ol-zoom-in, .ol-zoom-out {
+@media (min-width: theme('screens.sm')) {
+  .ol-zoom {
+    bottom: 94px !important;
+    display: flex;
+    animation: slideOut 300ms ease 0s 1 normal forwards;
+  }
+  .ol-zoom-open {
+    bottom: 94px !important;
+    display: flex;
+    animation: slideIn 500ms ease 0s 1 normal forwards;
+  }
+}
+
+@media (min-width: theme('screens.lg')) {
+  .ol-zoom {
+    bottom: 94px !important;
+    display: flex;
+    left: 554px !important;
+    animation: none;
+  }
+  .ol-zoom-open {
+    bottom: 94px !important;
+    display: flex;
+    left: 554px !important;
+    animation: none;
+  }
+}
+
+
+.ol-zoom-in, .ol-zoom-out, .ol-zoom-open-in, .ol-zoom-open-out {
   color: hsl(60, 100%, 95%) !important;
   background-color: hsl(210, 54%, 9%) !important;
   outline-offset: 0px !important;
@@ -156,17 +185,17 @@
   overflow: hidden !important;
 }
 
-.ol-zoom-in:hover, .ol-zoom-out:hover {
+.ol-zoom-in:hover, .ol-zoom-out:hover, .ol-zoom-open-in:hover, .ol-zoom-open-out:hover {
   background-color: hsl(60, 100%, 95%) !important;
   color: hsl(210, 54%, 9%) !important;
 }
 
-.ol-zoom-out {
+.ol-zoom-out, .ol-zoom-open-out {
   font-size: 32px !important;
   padding-bottom: 4px !important;
 }
 
-.ol-zoom-in {
+.ol-zoom-in, .ol-zoom-open-in {
   font-size: 24px !important;
 }
 
@@ -206,12 +235,12 @@
 
 .ol-swipe button::before {
   transform: rotate(135deg) !important;
-  left: -2px !important;
+  left: -2px;
 }
 
 .ol-swipe button::after {
   transform: rotate(-45deg) !important;
-  left: 2px !important;
+  left: 2px;
 }
 
 
@@ -219,4 +248,25 @@
   height: 15px;
   width: 15px;
   background-color: hsl(60, 100%, 95%);
+}
+
+@keyframes slideIn {
+	0% {
+		opacity: 0;
+    left: 16px;
+	}
+
+	100% {
+		opacity: 1;
+		left: 408px;
+	}
+}
+
+@keyframes slideOut {
+	0% {
+    left: 408px;
+	}
+	100% {
+    left: 16px;
+	}
 }
